requirements.txt
# run pip install -r requirements.txt to install the reqs!! :)

customtkinter==5.2.1
pyperclip==1.8.2
argon2-cffi==23.1.0
pycryptodome==3.20.0